Refactored code for IntTensor (#377)
