+++ UnityProject/Assets/OpenMined/Syft/Tensor/IntTensor.Ops.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 8f482334389dd114b804095c6b8d3c99
+timeCreated: 1516304646
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
